+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # vpploader changelog
 
+## 1.1.0 - 2025-05-17
+
+### Added
+
+- setOpacity function to VPPMesh class
+
+### Fixed
+
+- Color replacements not working properly
+
 ## 1.0.0 - 2025-02-22
 
 - Refactored into a module extending standard three.js Loader class